# Faster file stat