--- a/subprojects/native/src/test/groovy/org/gradle/internal/nativeintegration/filesystem/services/Jdk7FileMetadataAccessorTest.groovy
+++ b/subprojects/native/src/test/groovy/org/gradle/internal/nativeintegration/filesystem/services/Jdk7FileMetadataAccessorTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,12 +17,12 @@
 package org.gradle.internal.nativeintegration.filesystem.services
 
 import org.gradle.internal.nativeintegration.filesystem.FileMetadataAccessor
+import org.gradle.internal.nativeintegration.filesystem.jdk7.Jdk7FileMetadataAccessor
 import org.gradle.util.UsesNativeServices
 
 @UsesNativeServices
-class FallbackFileMetadataAccessorTest extends AbstractFileMetadataAccessorTest {
+class Jdk7FileMetadataAccessorTest extends AbstractFileMetadataAccessorTest {
     FileMetadataAccessor getAccessor() {
-        new FallbackFileMetadataAccessor()
-
+        new Jdk7FileMetadataAccessor()
     }
 }