While investigating performance regressions on `master`, I discovered that file stat (in particular because of annotation processor detection) was responsible for a large proportion of the performance regression. It is clearly visible in the following flame graph:

![selection_001](https://cloud.githubusercontent.com/assets/316357/21988332/2f5392ce-dc07-11e6-8a1f-f1aeaa24b67b.png)

This commit adds a new `Jdk7FileMetadataAccessor` which uses the JDK 7 NIO API to get file stats. We can therefore do the same in just 2 calls:

- one to check `exists`
- one to get all file stats (directory, length, last modified)

The difference is great:

![selection_002](https://cloud.githubusercontent.com/assets/316357/21988493/b02c7366-dc07-11e6-95ab-ecf193f5318e.png)

The regression goes from 10% slower to 4% slower (without the fixes I made in the other PR). We still need to make 2 calls, because `Files.exists` in JDK 7/8 is crazily expensive (because it internally throws an exception!). But this makes me wonder why `native-platform` would be slower than this, it makes no sense.