@melix 2017-01-17
The [commit above](https://github.com/gradle/gradle/pull/1183/commits/850c7d266e0c33c5a597900667b500cbcd7edde9) adds a JMH benchmark of the various implementations we have so far.

There are 4 different implementations:

- `Fallback` uses the good old `File` Java API, and consists of (subsequent) calls to `exists()`, `isDirectory()`, `lastModified()` and `length()`
- `NativePlatformBacked` is the version which uses @adammurdoch 's `native-platform` library
- `Jdk7` is the version which is added by this pull request. It consists of a first call to `File#exists()`, then, if it exists, uses nio's `DefaultFileAttributes` to get the rest of metadata (2 stat calls)
- for comparison, `nio` is the pure NIO version, which does everything on a single call

Here is the result on my laptop (Linux, ext4, SSD):

![selection_003](https://cloud.githubusercontent.com/assets/316357/22021509/16b7741c-dcbe-11e6-8294-21b81140f7b5.png)

What we can see is that the NIO version is significantly faster, **as long as the file exists**. Its performance is catastrophic if the file doesn't exist. Which is exactly why the `jdk7` version makes a first call to `File#exists` to avoid the internal exception that NIO triggers when the file is missing.

For existing files, native platform and jdk7 have barely the same performance, while the fallback version is significantly slower.

In short, given those results, I think it makes sense to use jdk7 by default under Linux. It would be nice to run this benchmark under OS X and Windows to see the difference. Eventually, it would be nice to find out why the JDK can be significantly faster than `native-platform`, because if we can remove this difference, then using `native-platform` everywhere would be the best solution.

/cc @oehme for the record