@melix 2017-01-17
Note that we could consider using the pure NIO implementation if we believe there won't be many "missing" files. Which I guess is adventurous as soon as you consider the "clean" case.