@melix 2017-01-17
Ah, I almost forgot I had a MacBook Pro at hand :D I'm going to run the same benchmark on it. BTW:

```
./gradlew native:jmhReport
```