@melix 2017-01-17
So here's the result for Mac OS:

![selection_004](https://cloud.githubusercontent.com/assets/316357/22022941/5e947aa4-dcc4-11e6-889f-599e6decf4af.png)

First we can see that it's almost twice as slow as under Linux, but `native-platform` is a clear winner here. So it's really different. Using the `jdk7` version would not be a bad choice compared to the fallback version, though, since it's always at least twice as fast, for the cases where the file actually exists.