@shipilev 2017-01-17
JDK bugs:
 https://bugs.openjdk.java.net/browse/JDK-8154077
 https://bugs.openjdk.java.net/browse/JDK-8153414