@melix 2017-01-17
Thanks for the links @shipilev :) It's good to know it's on the radar, yet we need solutions for everybody :)