@melix 2017-01-18
So here are the latest results, after rebasing on `release`, and using the `lotProjectDependencies` test that was failing.

With `jdk7` as the default accessor:

```
Speed Results for test project 'lotProjectDependencies' with tasks resolveDependencies: AWESOME! we're faster than 3.3 :D
Difference: 229.5 ms faster (229.5 ms), -6.53%, max regression: 123.938 ms
  Current Gradle median: 3.286 s min: 3.118 s, max: 3.591 s, se: 112.14 ms, sem: 15.859 ms
  > [3.306 s, 3.295 s, 3.285 s, 3.345 s, 3.431 s, 3.591 s, 3.23 s, 3.243 s, 3.315 s, 3.181 s, 3.22 s, 3.302 s, 3.204 s, 3.287 s, 3.194 s, 3.514 s, 3.413 s, 3.468 s, 3.424 s, 3.405 s, 3.238 s, 3.166 s, 3.238 s, 3.295 s, 3.137 s, 3.149 s, 3.126 s, 3.155 s, 3.118 s, 3.216 s, 3.132 s, 3.179 s, 3.143 s, 3.154 s, 3.177 s, 3.132 s, 3.236 s, 3.145 s, 3.195 s, 3.303 s, 3.357 s, 3.329 s, 3.339 s, 3.356 s, 3.386 s, 3.352 s, 3.335 s, 3.311 s, 3.411 s, 3.427 s]
  Gradle 3.3 median: 3.516 s min: 3.397 s, max: 3.99 s, se: 168.3 ms, sem: 23.801 ms
  > [3.751 s, 3.691 s, 3.6 s, 3.637 s, 3.605 s, 3.854 s, 3.566 s, 3.5 s, 3.593 s, 3.518 s, 3.564 s, 3.747 s, 3.591 s, 3.531 s, 3.511 s, 3.752 s, 3.925 s, 3.99 s, 3.781 s, 3.489 s, 3.479 s, 3.513 s, 3.47 s, 3.497 s, 3.416 s, 3.478 s, 3.487 s, 3.488 s, 3.441 s, 3.635 s, 3.884 s, 3.881 s, 3.936 s, 3.867 s, 3.49 s, 3.478 s, 3.474 s, 3.397 s, 3.411 s, 3.413 s, 3.463 s, 3.451 s, 3.412 s, 3.525 s, 3.513 s, 3.861 s, 3.825 s, 3.443 s, 3.397 s, 3.463 s]
```

Flame graph:

![jdk7](https://cloud.githubusercontent.com/assets/316357/22059102/c8c52ec6-dd6b-11e6-8c45-78168dddeadc.png)

Using `native-platform` as the default accessor:

```
Speed Results for test project 'lotProjectDependencies' with tasks resolveDependencies: AWESOME! we're faster than 3.3 :D
Difference: 177.5 ms faster (177.5 ms), -5.11%, max regression: 112.24 ms
  Current Gradle median: 3.294 s min: 3.192 s, max: 3.658 s, se: 112.26 ms, sem: 15.876 ms
  > [3.383 s, 3.424 s, 3.403 s, 3.488 s, 3.444 s, 3.433 s, 3.329 s, 3.318 s, 3.548 s, 3.536 s, 3.537 s, 3.658 s, 3.535 s, 3.294 s, 3.301 s, 3.422 s, 3.276 s, 3.378 s, 3.255 s, 3.264 s, 3.261 s, 3.251 s, 3.269 s, 3.318 s, 3.242 s, 3.258 s, 3.248 s, 3.209 s, 3.247 s, 3.237 s, 3.203 s, 3.342 s, 3.487 s, 3.49 s, 3.452 s, 3.436 s, 3.418 s, 3.22 s, 3.273 s, 3.284 s, 3.211 s, 3.31 s, 3.239 s, 3.29 s, 3.242 s, 3.199 s, 3.295 s, 3.192 s, 3.271 s, 3.27 s]
  Gradle 3.3 median: 3.472 s min: 3.372 s, max: 4.068 s, se: 141.71 ms, sem: 20.041 ms
  > [3.975 s, 4.068 s, 3.557 s, 3.638 s, 3.591 s, 3.645 s, 3.615 s, 3.472 s, 3.605 s, 3.443 s, 3.505 s, 3.699 s, 3.543 s, 3.826 s, 3.427 s, 3.509 s, 3.46 s, 3.531 s, 3.461 s, 3.431 s, 3.441 s, 3.473 s, 3.492 s, 3.508 s, 3.753 s, 3.432 s, 3.463 s, 3.462 s, 3.462 s, 3.479 s, 3.452 s, 3.693 s, 3.676 s, 3.453 s, 3.483 s, 3.399 s, 3.448 s, 3.433 s, 3.445 s, 3.472 s, 3.379 s, 3.461 s, 3.448 s, 3.423 s, 3.458 s, 3.655 s, 3.46 s, 3.375 s, 3.372 s, 3.501 s]
```

Flame graph:

![native-platform](https://cloud.githubusercontent.com/assets/316357/22059130/e9cf9214-dd6b-11e6-8e8d-1689eb673133.png)

There's barely a difference, but it does seem to confirm that `native-platform` is slower for this scenario. Good news is that we're faster than 3.3 now :)