@melix 2017-01-18
Here are results of the benchmarks executed on the CI server build agents.

Linux:

![ci-bench-linux](https://cloud.githubusercontent.com/assets/316357/22063698/73a10e46-dd80-11e6-8c4d-16a5b58d3342.png)

Windows:

![ci-bench-windows](https://cloud.githubusercontent.com/assets/316357/22063701/7972a320-dd80-11e6-8431-f8e2199c5ef6.png)

Apart from the fact that Windows is considerably slower (is that a surprise?), it seems to confirm that we should use:

- `native-platform` for OS X and Windows
- `jdk7` for Linux

And it would be interesting to know why `native-platform` is slower than the JDK version for Linux.