@adammurdoch 2017-01-18
Maybe also add a story to remind us to revisit whether we need the initial `File.exists()` call when running on Java 9 or later.